Move FilterInput
